added search functionality
Added a searchRepository, which can be implemented in the other reposotiries. Also created a search controller, which for now only has an endpoint for searching in jobs, by extending the jobService with a new seach function.

Search is implemented on the jobliste page. this should also be implemented on the product page